+++ mod.ts
@@ -2,7 +2,7 @@
 * # deno-logger
 
 [![NPM Version](https://img.shields.io/npm/v/@denodnt/logger.svg?style=flat)](https://www.npmjs.org/package/@denodnt/logger)
-[![JSR Version](https://jsr.io/badges/@deno-lib/logger)](https://jsr.io/@deno-lib/logger)
+[![JSR Version](https://jsr.io/badges/@deno-library/logger)](https://jsr.io/@deno-library/logger)
 
 Deno / NodeJS colorful logger colorful logger
 
@@ -13,7 +13,7 @@ For Deno usage refer to [deno-logger doc](https://deno.land/x/logger)
 ### console logger
 
 ```js
-import Logger from "jsr:@deno-lib/logger@1.1.6/logger";
+import { Logger } from "jsr:@deno-library/logger@1.1.7";
 
 const logger = new Logger();
 
@@ -25,7 +25,7 @@ logger.error("i am from consoleLogger", new Error("test"));
 ### file and console logger
 
 ```js
-import Logger from "jsr:@deno-lib/logger@1.1.6/logger";
+import { Logger } from "jsr:@deno-library/logger@1.1.7";
 
 const logger = new Logger();
 
@@ -45,7 +45,7 @@ logger.error("i am from fileLogger", new Error("test"));
 ### file logger only
 
 ```js
-import Logger from "jsr:@deno-lib/logger@1.1.6/logger";
+import { Logger } from "jsr:@deno-library/logger@1.1.7";
 
 const logger = new Logger();
 await logger.initFileLogger("../log");
@@ -71,7 +71,7 @@ interface fileLoggerOptions {
 example
 
 ```js
-import Logger from "jsr:@deno-lib/logger@1.1.6/logger";
+import { Logger } from "jsr:@deno-library/logger@1.1.7";
 const logger = new Logger();
 
 // cut by day
@@ -159,7 +159,7 @@ await logger.initFileLogger("../log", {
 ## disableConsole and enableConsole
 
 ```js
-import Logger from "jsr:@deno-lib/logger@1.1.6/logger";
+import { Logger } from "jsr:@deno-library/logger@1.1.7";
 
 const logger = new Logger();
 
@@ -176,7 +176,7 @@ logger.info("console enabled, you can see me");
 ## disableFile and enableFile
 
 ```js
-import Logger from "jsr:@deno-lib/logger@1.1.6/logger";
+import { Logger } from "jsr:@deno-library/logger@1.1.7";
 
 const logger = new Logger();
 await logger.initFileLogger("../log");
@@ -203,7 +203,7 @@ example:
 3. fileLogger, consoleLogger => disable => enable => fileLogger, consoleLogger
 
 ```js
-import Logger from "jsr:@deno-lib/logger@1.1.6/logger";
+import { Logger } from "jsr:@deno-library/logger@1.1.7";
 
 const logger = new Logger();
 await logger.initFileLogger("../log");
@@ -250,9 +250,9 @@ await build({
   },
   // map your favorite deno logger to its npm port.
   mappings: {
-    "jsr:@deno-lib/logger@1.1.6/logger": {
+    "jsr:@deno-library/logger@1.1.7/logger": {
       name: "@denodnt/logger",
-      version: "1.1.6",
+      version: "1.1.7",
       peerDependency: false,
     },
   },