+++ changelog.md
@@ -3,6 +3,7 @@
 ### v1.1.7 - 2024.11.11
 
 - [Update the used deno API](https://github.com/deno-library/logger/pull/17)
+- move from `deno-lib/logger` to `deno-library/logger`
 
 ### v1.1.6 - 2024.05.17